+++ modules/azuredevops_serviceendpoint/outputs.tf
@@ -23,12 +23,12 @@ output "policy_ids" {
     branch_min_reviewers      = { for key, policy in azuredevops_branch_policy_min_reviewers.policy : key => policy.id }
     branch_status_check       = { for key, policy in azuredevops_branch_policy_status_check.policy : key => policy.id }
     branch_work_item_linking  = { for key, policy in azuredevops_branch_policy_work_item_linking.policy : key => policy.id }
-    repo_author_email_pattern = { for key, policy in azuredevops_repository_policy_author_email_pattern.policy : key => policy.id }
-    repo_case_enforcement     = { for key, policy in azuredevops_repository_policy_case_enforcement.policy : key => policy.id }
-    repo_check_credentials    = { for key, policy in azuredevops_repository_policy_check_credentials.policy : key => policy.id }
-    repo_file_path_pattern    = { for key, policy in azuredevops_repository_policy_file_path_pattern.policy : key => policy.id }
-    repo_max_file_size        = { for key, policy in azuredevops_repository_policy_max_file_size.policy : key => policy.id }
-    repo_max_path_length      = { for key, policy in azuredevops_repository_policy_max_path_length.policy : key => policy.id }
-    repo_reserved_names       = { for key, policy in azuredevops_repository_policy_reserved_names.policy : key => policy.id }
+    repo_author_email_pattern = { for key, policy in azuredevops_serviceendpoint_policy_author_email_pattern.policy : key => policy.id }
+    repo_case_enforcement     = { for key, policy in azuredevops_serviceendpoint_policy_case_enforcement.policy : key => policy.id }
+    repo_check_credentials    = { for key, policy in azuredevops_serviceendpoint_policy_check_credentials.policy : key => policy.id }
+    repo_file_path_pattern    = { for key, policy in azuredevops_serviceendpoint_policy_file_path_pattern.policy : key => policy.id }
+    repo_max_file_size        = { for key, policy in azuredevops_serviceendpoint_policy_max_file_size.policy : key => policy.id }
+    repo_max_path_length      = { for key, policy in azuredevops_serviceendpoint_policy_max_path_length.policy : key => policy.id }
+    repo_reserved_names       = { for key, policy in azuredevops_serviceendpoint_policy_reserved_names.policy : key => policy.id }
   }
 }